requirements.txt
accelerate==0.27.2
black==24.4.2
datasets==2.17.1
jiwer==3.0.3
librosa==0.10.1
openai-whisper==20231117
pylint==3.0.3
soundfile==0.12.1
toml==0.10.2
transformers==4.37.1
whisper_timestamped==1.15.1
